Extract out an `options.rs` module
